pyproject.toml
[project]
name = "rossmann-mlops"
version = "0.1.0"
description = "MLOps pipeline for Rossmann sales forecasting"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.0",
    "joblib>=1.4.2",
    "numpy>=2.1.0",
    "pandas>=2.2.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.5.0",
    "uvicorn>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
]
data = [
    "dvc>=3.67.1",
    "dvc-gdrive>=3.0.1",
    "dvc-s3>=3.3.0",
]

[project.scripts]
run-pipeline = "rossmann_mlops.train:main"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
